GS: Move dump file loading to core
